refactor: create the MP3 parser
the id3v2 is now only used for the tag parsing
